requirements.txt
# Dependencies
asyncclick
dagger-io
